data files and script upload
Datafiles associated with the manuscript. The folder geum_physiology_datasheets contains the clean datafiles that were used in the scripts for analysis and figure creation (metadata for these files is in the metadata subfolder). geum_physiology_r_scripts is where the scripts for analysis are contained, with the bulk of analysis preserved in an .rmd file. RAW_physiology_datasheets contains all the raw datafiles.